Extract an AbstractReadFilterPluginDescriptor
